refactor(#3038): extract submit_review_decision domain logic into services::review_decision (thin HTTP handler)
