+++ ARCHITECTURE.md
@@ -665,6 +665,7 @@ src/
 │   ├── qwen_tmux_wrapper.rs
 │   ├── remote_stub.rs
 │   ├── retrospectives.rs
+│   ├── review_decision.rs
 │   ├── service_error.rs
 │   ├── session_backend.rs
 │   ├── session_forwarding.rs
