+++ src/services/mod.rs
@@ -59,6 +59,7 @@ pub mod qwen;
 pub mod qwen_tmux_wrapper;
 pub mod remote_stub;
 pub mod retrospectives;
+pub mod review_decision;
 pub mod routines;
 pub mod service_error;
 pub mod session_backend;
